refactor: rename to ViewItemGrid and EditItemGrid
