+++ src/components/Dashboard/__tests__/ViewDashboard.spec.js
@@ -12,7 +12,7 @@ jest.mock('../../ControlBar/DashboardsBar', () => 'DashboardsBar')
 jest.mock('../DashboardVerticalOffset', () => 'DashboardVerticalOffset')
 jest.mock('../../TitleBar/TitleBar', () => 'TitleBar')
 jest.mock('../../FilterBar/FilterBar', () => 'FilterBar')
-jest.mock('../../ItemGrid/ItemGrid', () => 'ItemGrid')
+jest.mock('../../ItemGrid/ViewItemGrid', () => 'ViewItemGrid')
 
 describe('ViewDashboard', () => {
     let props
